Move Genre components into their own folder
